move test to parent component
